requirements.txt
absl-py==2.4.0
APScheduler==3.10.4
astunparse==1.6.3
blinker==1.9.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
colorhash==2.3.0
configparser==7.2.0
cryptography==46.0.7
Flask==3.1.3
Flask-MonitoringDashboard @ file:///home/yacine/DEEP-I/E5-1/Flask-MonitoringDashboard
flatbuffers==25.12.19
gast==0.7.0
google-api-core==2.30.3
google-api-python-client==2.194.0
google-auth==2.49.2
google-auth-httplib2==0.3.1
google-auth-oauthlib==1.3.1
google-pasta==0.2.0
googleapis-common-protos==1.74.0
greenlet==3.4.0
grpcio==1.80.0
h5py==3.14.0
httplib2==0.31.2
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
keras==3.14.0
libclang==18.1.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ml_dtypes==0.5.4
namex==0.1.0
narwhals==2.19.0
numpy==2.4.4
oauthlib==3.3.1
opt_einsum==3.4.0
optree==0.19.0
packaging==26.0
pandas==3.0.2
plotly==6.7.0
proto-plus==1.27.2
protobuf==7.34.1
psutil==7.2.2
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
Pygments==2.20.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-http-client==3.3.7
pytz==2026.1.post1
requests==2.33.1
requests-oauthlib==2.0.0
rich==15.0.0
scipy==1.17.1
sendgrid==6.12.5
six==1.17.0
SQLAlchemy==2.0.49
tensorflow==2.21.0
termcolor==3.3.0
typing_extensions==4.15.0
tzlocal==2.0.0
uritemplate==4.2.0
urllib3==2.6.3
Werkzeug==3.1.8
wrapt==2.1.2
